refactor program
