+++ programs/sunrise-stake/src/instructions.rs
@@ -1,0 +1,23 @@
+mod claim_unstake_ticket;
+mod create_metadata;
+mod deposit;
+mod deposit_stake_account;
+mod extract_to_treasury;
+mod liquid_unstake;
+mod order_unstake;
+mod register_state;
+mod trigger_pool_rebalance;
+mod update_metadata;
+mod update_state;
+
+pub use claim_unstake_ticket::*;
+pub use create_metadata::*;
+pub use deposit::*;
+pub use deposit_stake_account::*;
+pub use extract_to_treasury::*;
+pub use liquid_unstake::*;
+pub use order_unstake::*;
+pub use register_state::*;
+pub use trigger_pool_rebalance::*;
+pub use update_metadata::*;
+pub use update_state::*;
