+++ programs/sunrise-stake/src/lib.rs
@@ -2,32 +2,20 @@
 mod sunrise_spl;
 mod utils;
 
-use crate::utils::marinade::ClaimUnstakeTicketProperties;
-use crate::utils::metaplex::{create_metadata_account, update_metadata_account};
-use crate::utils::seeds::*;
-use crate::utils::token::{create_mint, mint_to};
-use crate::utils::{
-    marinade,
-    marinade::{
-        amount_to_be_deposited_in_liq_pool, calc_lamports_from_msol_amount,
-        calc_msol_from_lamports, calculate_extractable_yield, calculate_pool_balance_amounts,
-    },
-    system,
-    token::{burn, create_token_account},
-};
+use crate::utils::metaplex::update_metadata_account;
 use anchor_lang::prelude::borsh::BorshDeserialize;
 use anchor_lang::prelude::*;
-use anchor_lang::solana_program::{
-    program::invoke_signed, program_option::COption, system_instruction::transfer,
-};
-use anchor_spl::associated_token::AssociatedToken;
-use anchor_spl::token::{Mint, Token, TokenAccount};
-use marinade_cpi::program::MarinadeFinance;
-use marinade_cpi::{State as MarinadeState, TicketAccountData as MarinadeTicketAccount};
+use anchor_spl::token::Mint;
 use sunrise_spl::*;
 
-use anchor_lang::AccountsClose;
-use std::ops::Deref;
+pub mod state;
+use crate::state::*;
+
+pub mod error;
+use crate::error::ErrorCode;
+
+pub mod instructions;
+use instructions::*;
 
 declare_id!("sunzv8N3A8dRHwUBvxgRDEbWKk8t7yiHR4FLRgFsTX6");
 
@@ -36,159 +24,26 @@ pub mod sunrise_stake {
     use super::*;
 
     pub fn deposit(ctx: Context<Deposit>, lamports: u64) -> Result<()> {
-        msg!("Checking liq_pool pool balance");
-        let to_deposit_in_liq_pool = amount_to_be_deposited_in_liq_pool(ctx.accounts, lamports)?;
-        let to_stake = lamports - to_deposit_in_liq_pool;
-
-        if to_deposit_in_liq_pool > 0 {
-            msg!("Depositing {} in liq_pool pool", to_deposit_in_liq_pool);
-            let accounts = ctx.accounts.deref().into();
-            marinade::add_liquidity(&accounts, to_deposit_in_liq_pool)?;
-        }
-
-        if to_stake > 0 {
-            msg!("Staking {}", to_stake);
-            marinade::deposit(ctx.accounts, to_stake)?;
-        }
-
-        msg!("Mint {} GSOL", lamports);
-        mint_to(
-            lamports,
-            &ctx.accounts.gsol_mint.to_account_info(),
-            &ctx.accounts.gsol_mint_authority.to_account_info(),
-            &ctx.accounts.mint_gsol_to.to_account_info(),
-            &ctx.accounts.token_program.to_account_info(),
-            &ctx.accounts.state,
-        )?;
-        let state = &mut ctx.accounts.state;
-        state.marinade_minted_gsol = state.marinade_minted_gsol.checked_add(lamports).unwrap();
-        Ok(())
+        deposit_handler(ctx, lamports)
     }
 
     pub fn deposit_stake_account(
         ctx: Context<DepositStakeAccount>,
         validator_index: u32,
     ) -> Result<()> {
-        let lamports = marinade::get_delegated_stake_amount(&ctx.accounts.stake_account)?;
-
-        msg!("Depositing stake account");
-        marinade::deposit_stake_account(ctx.accounts, validator_index)?;
-
-        msg!("Mint {} GSOL", lamports);
-        mint_to(
-            lamports,
-            &ctx.accounts.gsol_mint.to_account_info(),
-            &ctx.accounts.gsol_mint_authority.to_account_info(),
-            &ctx.accounts.mint_gsol_to.to_account_info(),
-            &ctx.accounts.token_program.to_account_info(),
-            &ctx.accounts.state,
-        )?;
-        let state = &mut ctx.accounts.state;
-        state.marinade_minted_gsol = state.marinade_minted_gsol.checked_add(lamports).unwrap();
-        Ok(())
+        deposit_stake_account_handler(ctx, validator_index)
     }
 
     pub fn order_unstake(ctx: Context<OrderUnstake>, lamports: u64) -> Result<()> {
-        let msol_lamports =
-            calc_msol_from_lamports(ctx.accounts.marinade_state.as_ref(), lamports)?;
-
-        let lamports_converted =
-            calc_lamports_from_msol_amount(ctx.accounts.marinade_state.as_ref(), msol_lamports)?;
-
-        msg!(
-            "Ordering unstake of {} MSOL (in lamports {}, out lamports {})",
-            msol_lamports,
-            lamports,
-            lamports_converted
-        );
-        let accounts = ctx.accounts.deref().into();
-        marinade::order_unstake(&accounts, msol_lamports)?;
-
-        msg!("Ticket beneficiary {}", ctx.accounts.get_msol_from.owner);
-
-        ctx.accounts.sunrise_ticket_account.state_address = ctx.accounts.state.key();
-        ctx.accounts.sunrise_ticket_account.marinade_ticket_account =
-            ctx.accounts.new_ticket_account.key();
-        ctx.accounts.sunrise_ticket_account.beneficiary =
-            ctx.accounts.gsol_token_account_authority.key();
-
-        msg!("Burn GSol");
-        burn(
-            lamports,
-            &ctx.accounts.gsol_mint.to_account_info(),
-            &ctx.accounts.gsol_token_account_authority.to_account_info(),
-            &ctx.accounts.gsol_token_account.to_account_info(),
-            &ctx.accounts.token_program.to_account_info(),
-        )?;
-
-        let state = &mut ctx.accounts.state;
-        state.marinade_minted_gsol = state.marinade_minted_gsol.checked_sub(lamports).unwrap();
-
-        Ok(())
+        order_unstake_handler(ctx, lamports)
     }
 
     pub fn claim_unstake_ticket(ctx: Context<ClaimUnstakeTicket>) -> Result<()> {
-        let accounts = ctx.accounts.deref().into();
-        marinade::claim_unstake_ticket(&accounts)?;
-
-        // transfer the released SOL to the beneficiary
-        let lamports = ctx.accounts.marinade_ticket_account.lamports_amount;
-        let ix = transfer(
-            &ctx.accounts.msol_authority.key(),
-            &ctx.accounts.transfer_sol_to.key(),
-            lamports,
-        );
-        let bump = &[ctx.accounts.state.msol_authority_bump][..];
-        let state_address = ctx.accounts.state.key();
-        let seeds = &[state_address.as_ref(), MSOL_ACCOUNT, bump][..];
-        invoke_signed(
-            &ix,
-            &[
-                ctx.accounts.msol_authority.to_account_info(),
-                ctx.accounts.transfer_sol_to.to_account_info(),
-            ],
-            &[seeds],
-        )?;
-
-        Ok(())
+        claim_unstake_ticket_handler(ctx)
     }
 
     pub fn liquid_unstake(ctx: Context<LiquidUnstake>, lamports: u64) -> Result<()> {
-        msg!("Checking liq_pool pool balance");
-        let calculate_balance_props = ctx.accounts.deref().into();
-        let amounts = calculate_pool_balance_amounts(&calculate_balance_props, lamports)?;
-
-        if amounts.amount_to_withdraw_from_liq_pool.liq_pool_token > 0 {
-            marinade::remove_liquidity(
-                ctx.accounts,
-                amounts.amount_to_withdraw_from_liq_pool.liq_pool_token,
-            )?;
-        }
-
-        if amounts.amount_to_liquid_unstake > 0 {
-            let msol_lamports = calc_msol_from_lamports(
-                ctx.accounts.marinade_state.as_ref(),
-                amounts.amount_to_liquid_unstake,
-            )?;
-
-            msg!("Unstaking {} msol lamports", msol_lamports);
-            let accounts = ctx.accounts.deref().into();
-            marinade::unstake(&accounts, msol_lamports)?;
-        }
-
-        msg!("Burn GSol");
-        burn(
-            lamports,
-            &ctx.accounts.gsol_mint.to_account_info(),
-            &ctx.accounts.gsol_token_account_authority,
-            &ctx.accounts.gsol_token_account.to_account_info(),
-            &ctx.accounts.token_program.to_account_info(),
-        )?;
-
-        let state = &mut ctx.accounts.state;
-        state.marinade_minted_gsol = state.marinade_minted_gsol.checked_sub(lamports).unwrap();
-
-        Ok(())
+        liquid_unstake_handler(ctx, lamports)
     }
 
     pub fn trigger_pool_rebalance<'info>(
@@ -198,137 +53,17 @@ pub mod sunrise_stake {
         order_unstake_ticket_account_bump: u8,
         _previous_order_unstake_ticket_management_account_bump: u8,
     ) -> Result<()> {
-        // check the ticket info is correct
-        require_eq!(ctx.accounts.clock.epoch, epoch);
-        require_eq!(
-            ctx.accounts.order_unstake_ticket_management_account.tickets + 1,
-            index
-        );
-
-        let management_account = &mut ctx.accounts.order_unstake_ticket_management_account;
-        if management_account.epoch == 0 {
-            // set up management account
-            management_account.epoch = epoch;
-            management_account.tickets = 0;
-        } else {
-            // check the epoch is correct
-            require_eq!(management_account.epoch, epoch);
-        }
-
-        let mut claimed_lamports = 0;
-        let mut props: ClaimUnstakeTicketProperties = ctx.accounts.deref().into();
-        // All remaining accounts are previous epoch tickets that are now ready to be claimed.
-        msg!("Tickets to claim: {}", ctx.remaining_accounts.len());
-        for ticket in ctx.remaining_accounts.iter() {
-            let ticket_account = TicketAccountData::try_from_slice(&ticket.data.borrow_mut())?;
-            claimed_lamports += ticket_account.lamports_amount;
-
-            msg!(
-                "Claiming ticket {} with value {}",
-                ticket.key(),
-                ticket_account.lamports_amount
-            );
-
-            props.ticket_account = ticket.to_account_info();
-
-            marinade::claim_unstake_ticket(&props)?;
-        }
-
-        if claimed_lamports > 0 {
-            msg!(
-                "Claimed {} lamports from tickets - depositing into liquidity pool",
-                claimed_lamports
-            );
-
-            let add_liquidity_props = ctx.accounts.deref().into();
-            let lamports_to_deposit = ctx.accounts.get_msol_from_authority.try_lamports()?;
-
-            msg!(
-                "Current balance of msol token authority {}, claiming amount {}",
-                lamports_to_deposit,
-                claimed_lamports
-            );
-            marinade::add_liquidity_from_pda(&add_liquidity_props, claimed_lamports)?;
-        }
-
-        msg!(
-            "Before reload {}",
-            ctx.accounts.liq_pool_token_account.amount
-        );
-        ctx.accounts.liq_pool_token_account.reload()?;
-        msg!("Reloaded liq_pool_token_amount to reflect any added liquidity");
-        msg!(
-            "after reload {}",
-            ctx.accounts.liq_pool_token_account.amount
-        );
-
-        msg!("Checking liq_pool pool balance");
-        let calculate_pool_balance_props = ctx.accounts.deref().into();
-        let amounts = calculate_pool_balance_amounts(&calculate_pool_balance_props, 0)?;
-
-        if amounts.amount_to_order_delayed_unstake > 0 {
-            let msol_lamports = calc_msol_from_lamports(
-                ctx.accounts.marinade_state.as_ref(),
-                amounts.amount_to_order_delayed_unstake,
-            )?;
-
-            // TODO move to just using init
-            msg!("Creating order unstake ticket account");
-            let create_ticket_props = ctx.accounts.deref().into();
-            system::create_order_unstake_ticket_account(
-                &create_ticket_props,
-                order_unstake_ticket_account_bump,
-                index,
-            )?;
-
-            msg!(
-                "Ordering a delayed unstake of {} msol lamports",
-                msol_lamports
-            );
-            let order_unstake_props = ctx.accounts.deref().into();
-            marinade::order_unstake(&order_unstake_props, msol_lamports)?;
-
-            // updating the internal record of delayed unstakes ordered.
-            ctx.accounts
-                .order_unstake_ticket_management_account
-                .add_ticket(
-                    amounts.amount_to_order_delayed_unstake,
-                    ctx.accounts.clock.epoch,
-                )?;
-        }
-
-        // WARNING - this must happen _after_ the order unstake ticket account is created
-        // otherwise the transaction fails with "sum of account balances before and after instruction do not match"
-        // TODO check here that all tickets for the previous epoch are now closed
-        // Close the previous epoch's ticket management account
-        // and pass the rent to the tx payer, which compensates them for opening a new one
-        msg!("Closing previous epoch's ticket management account");
-        ctx.accounts
-            .previous_order_unstake_ticket_management_account
-            .close(ctx.accounts.payer.to_account_info())?;
-
-        Ok(())
+        trigger_pool_rebalance_handler(
+            ctx,
+            epoch,
+            index,
+            order_unstake_ticket_account_bump,
+            _previous_order_unstake_ticket_management_account_bump,
+        )
     }
 
     pub fn extract_to_treasury(ctx: Context<ExtractToTreasury>) -> Result<()> {
-        // TODO at present, this withdraws all msol yield. In future, we should be able to choose how much to withdraw
-        let extractable_yield = calculate_extractable_yield(
-            ctx.accounts,
-            &ctx.accounts.get_msol_from,
-            &ctx.accounts.get_bsol_from,
-            &ctx.accounts.gsol_mint,
-        )?;
-
-        let extractable_yield_msol =
-            calc_msol_from_lamports(ctx.accounts.marinade_state.as_ref(), extractable_yield)?;
-
-        // TODO later change to use "slow unstake" rather than incur liq pool fees
-
-        msg!("Withdrawing {} msol to treasury", extractable_yield);
-        let accounts = ctx.accounts.deref().into();
-        marinade::unstake(&accounts, extractable_yield_msol)?;
-
-        Ok(())
+        extract_to_treasury_handler(ctx)
     }
 
     //////////////////////////////////////////
@@ -355,133 +90,11 @@ pub mod sunrise_stake {
     // ADMIN FUNCTIONS
     ////////////////////////////
     pub fn register_state(ctx: Context<RegisterState>, state: StateInput) -> Result<()> {
-        let state_account = &mut ctx.accounts.state;
-        state_account.set_values(&state, &ctx.accounts.mint.key(), 0, 0);
-
-        // create the gsol mint
-        let gsol_mint_authority = Pubkey::create_program_address(
-            &[
-                &state_account.key().to_bytes(),
-                GSOL_MINT_AUTHORITY,
-                &[state_account.gsol_mint_authority_bump],
-            ],
-            ctx.program_id,
-        )
-        .unwrap();
-        create_mint(
-            &ctx.accounts.payer,
-            &ctx.accounts.mint.to_account_info(),
-            &gsol_mint_authority,
-            &ctx.accounts.system_program,
-            &ctx.accounts.token_program,
-            &ctx.accounts.rent.to_account_info(),
-        )?;
-
-        // create msol token account
-        // Note - the relationship between msol_mint and marinade_state is not verified here
-        // Specifically, the marinade_state is not passed into the register function as an account.
-        // This simplifies the registration code, but if it is registered incorrectly, deposits will fail.
-        create_token_account(
-            &ctx.accounts.payer,
-            &ctx.accounts.msol_token_account,
-            &ctx.accounts.msol_mint,
-            &ctx.accounts.msol_token_account_authority,
-            &ctx.accounts.system_program,
-            &ctx.accounts.token_program,
-            &ctx.accounts.associated_token_program,
-        )?;
-
-        // create marinade msol/sol liquidity pool token account
-        // the same token account authority PDA is used for the
-        // msol token account and the liquidity pool token account
-        create_token_account(
-            &ctx.accounts.payer,
-            &ctx.accounts.liq_pool_token_account,
-            &ctx.accounts.liq_pool_mint,
-            &ctx.accounts.msol_token_account_authority,
-            &ctx.accounts.system_program,
-            &ctx.accounts.token_program,
-            &ctx.accounts.associated_token_program,
-        )?;
-
-        // create bsol token account
-        // Note - the relationship between bsol_mint and blaze_state is not verified here
-        // Specifically, the blaze_state is not passed into the register function as an account.
-        // This simplifies the registration code, but if it is registered incorrectly, deposits will fail.
-        create_token_account(
-            &ctx.accounts.payer,
-            &ctx.accounts.bsol_token_account,
-            &ctx.accounts.bsol_mint,
-            &ctx.accounts.bsol_token_account_authority,
-            &ctx.accounts.system_program,
-            &ctx.accounts.token_program,
-            &ctx.accounts.associated_token_program,
-        )?;
-
-        Ok(())
+        register_state_handler(ctx, state)
     }
 
     pub fn update_state(ctx: Context<UpdateState>, state: StateInput) -> Result<()> {
-        // Check the liq_pool_proportion does not exceed 100%
-        require_gte!(100, state.liq_pool_proportion);
-
-        let state_account = &mut ctx.accounts.state;
-        let gsol_mint = state_account.gsol_mint;
-
-        let marinade_minted_gsol = state_account.marinade_minted_gsol;
-        let blaze_minted_gsol = state_account.blaze_minted_gsol;
-
-        state_account.set_values(&state, &gsol_mint, marinade_minted_gsol, blaze_minted_gsol);
-
-        // Create any token accounts not yet created
-        if *ctx.accounts.msol_token_account.owner != ctx.accounts.token_program.key() {
-            // create msol token account
-            // Note - the relationship between msol_mint and marinade_state is not verified here
-            // Specifically, the marinade_state is not passed into the register function as an account.
-            // This simplifies the registration code, but if it is registered incorrectly, deposits will fail.
-            create_token_account(
-                &ctx.accounts.payer,
-                &ctx.accounts.msol_token_account,
-                &ctx.accounts.msol_mint,
-                &ctx.accounts.msol_token_account_authority,
-                &ctx.accounts.system_program,
-                &ctx.accounts.token_program,
-                &ctx.accounts.associated_token_program,
-            )?;
-        }
-
-        if *ctx.accounts.liq_pool_token_account.owner != ctx.accounts.token_program.key() {
-            // create marinade msol/sol liquidity pool token account
-            // the same token account authority PDA is used for the
-            // msol token account and the liquidity pool token account
-            create_token_account(
-                &ctx.accounts.payer,
-                &ctx.accounts.liq_pool_token_account,
-                &ctx.accounts.liq_pool_mint,
-                &ctx.accounts.msol_token_account_authority,
-                &ctx.accounts.system_program,
-                &ctx.accounts.token_program,
-                &ctx.accounts.associated_token_program,
-            )?;
-        }
-
-        if *ctx.accounts.bsol_token_account.owner != ctx.accounts.token_program.key() {
-            // create bsol token account
-            // Note - the relationship between bsol_mint and blaze_state is not verified here
-            // Specifically, the blaze_state is not passed into the register function as an account.
-            // This simplifies the registration code, but if it is registered incorrectly, deposits will fail.
-            create_token_account(
-                &ctx.accounts.payer,
-                &ctx.accounts.bsol_token_account,
-                &ctx.accounts.bsol_mint,
-                &ctx.accounts.bsol_token_account_authority,
-                &ctx.accounts.system_program,
-                &ctx.accounts.token_program,
-                &ctx.accounts.associated_token_program,
-            )?;
-        }
-
-        Ok(())
+        update_state_handler(ctx, state)
     }
 
     // TODO this is only used during development, to add features to the state account
@@ -491,15 +104,13 @@ pub mod sunrise_stake {
         Ok(())
     }
 
-    // used once to create token metadata for gSOL
     pub fn create_metadata(
         ctx: Context<CreateMetadata>,
         uri: String,
         name: String,
         symbol: String,
     ) -> Result<()> {
-        msg!("Create Metadata for gSol");
-        create_metadata_account(ctx.accounts, uri, name, symbol)
+        create_metadata_handler(ctx, uri, name, symbol)
     }
 
     // used once to create token metadata for gSOL
@@ -511,55 +122,6 @@ pub mod sunrise_stake {
     ) -> Result<()> {
         msg!("Update Metadata for gSol");
         update_metadata_account(ctx.accounts, uri, name, symbol)
-    }
-}
-
-#[account]
-pub struct State {
-    pub marinade_state: Pubkey,
-
-    pub update_authority: Pubkey,
-    pub gsol_mint: Pubkey,
-    pub treasury: Pubkey,
-    pub gsol_mint_authority_bump: u8,
-    pub msol_authority_bump: u8,
-
-    /// 0-100 - The proportion of the total staked SOL that should be in the
-    /// liquidity pool.
-    pub liq_pool_proportion: u8,
-    /// 0-100 - If unstaking would result in the proportion of SOL in the
-    /// liquidity pool dropping below this value, trigger an delayed unstake
-    /// for the difference
-    pub liq_pool_min_proportion: u8,
-
-    pub blaze_state: Pubkey,
-    pub marinade_minted_gsol: u64,
-    pub blaze_minted_gsol: u64,
-    pub bsol_authority_bump: u8,
-}
-
-impl State {
-    const SPACE: usize = 32 + 32 + 32 + 32 + 1 + 1 + 1 + 1 + 32 + 8 + 8 + 1 + 8 /* DISCRIMINATOR */ ;
-
-    pub fn set_values(
-        &mut self,
-        input: &StateInput,
-        gsol_mint: &Pubkey,
-        marinade_minted_gsol: u64,
-        blaze_minted_gsol: u64,
-    ) {
-        self.marinade_state = input.marinade_state;
-        self.blaze_state = input.blaze_state;
-        self.update_authority = input.update_authority;
-        self.gsol_mint_authority_bump = input.gsol_mint_authority_bump;
-        self.msol_authority_bump = input.msol_authority_bump;
-        self.bsol_authority_bump = input.bsol_authority_bump;
-        self.treasury = input.treasury;
-        self.gsol_mint = *gsol_mint;
-        self.liq_pool_proportion = input.liq_pool_proportion;
-        self.liq_pool_min_proportion = input.liq_pool_min_proportion;
-        self.marinade_minted_gsol = marinade_minted_gsol;
-        self.blaze_minted_gsol = blaze_minted_gsol;
     }
 }
 
@@ -582,836 +144,4 @@ pub fn check_mint_supply(state: &State, gsol_mint: &Account<Mint>) -> Result<()>
         ErrorCode::UnexpectedMintSupply
     );
     Ok(())
-}
-
-/// Maps a marinade ticket account to a GSOL token holder
-#[account]
-pub struct SunriseTicketAccount {
-    pub state_address: Pubkey, // instance of sunrise state this ticket belongs to
-    pub marinade_ticket_account: Pubkey,
-    pub beneficiary: Pubkey,
-}
-impl SunriseTicketAccount {
-    const SPACE: usize = 32 + 32 + 32 + 8 /* DISCRIMINATOR */ ;
-}
-
-#[account]
-pub struct OrderUnstakeTicketManagementAccount {
-    pub state_address: Pubkey,
-    pub epoch: u64,
-    pub tickets: u64,
-    pub total_ordered_lamports: u64,
-}
-impl OrderUnstakeTicketManagementAccount {
-    const SPACE: usize = 32 + 8 + 8 + 8 + 8 /* DISCRIMINATOR */ ;
-
-    pub fn add_ticket(&mut self, ticket_amount_lamports: u64, epoch: u64) -> Result<()> {
-        if self.epoch == 0 {
-            self.epoch = epoch;
-        } else {
-            require_eq!(
-                self.epoch,
-                epoch,
-                ErrorCode::InvalidOrderUnstakeManagementAccount
-            );
-        }
-        self.tickets += 1;
-        self.total_ordered_lamports += ticket_amount_lamports;
-        Ok(())
-    }
-}
-
-#[derive(AnchorSerialize, AnchorDeserialize, Clone)]
-pub struct StateInput {
-    pub marinade_state: Pubkey,
-    pub blaze_state: Pubkey,
-    pub update_authority: Pubkey,
-    pub treasury: Pubkey,
-    pub gsol_mint_authority_bump: u8,
-    pub msol_authority_bump: u8,
-    pub bsol_authority_bump: u8,
-    pub liq_pool_proportion: u8,
-    pub liq_pool_min_proportion: u8,
-}
-
-#[derive(Accounts)]
-#[instruction(state_in: StateInput)]
-pub struct RegisterState<'info> {
-    #[account(init, space = State::SPACE, payer = payer)]
-    pub state: Account<'info, State>,
-
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    #[account(mut)]
-    pub mint: Signer<'info>,
-
-    #[account()]
-    pub msol_mint: Box<Account<'info, Mint>>,
-    #[account()]
-    pub bsol_mint: Box<Account<'info, Mint>>,
-
-    /// Must be a PDA, but otherwise owned by the system account ie not initialised with data
-    #[account(
-    seeds = [state.key().as_ref(), MSOL_ACCOUNT],
-    bump = state_in.msol_authority_bump
-    )]
-    pub msol_token_account_authority: SystemAccount<'info>,
-
-    #[account(mut)]
-    pub msol_token_account: SystemAccount<'info>,
-
-    #[account()]
-    pub liq_pool_mint: Box<Account<'info, Mint>>,
-
-    #[account(mut)]
-    pub liq_pool_token_account: SystemAccount<'info>,
-
-    #[account(
-    seeds = [state.key().as_ref(), BSOL_ACCOUNT],
-    bump = state_in.bsol_authority_bump
-    )]
-    pub bsol_token_account_authority: SystemAccount<'info>,
-
-    /// CHECK: Checked by AssociatedTokenAccount program
-    #[account(mut)]
-    pub bsol_token_account: SystemAccount<'info>,
-
-    pub system_program: Program<'info, System>,
-    pub token_program: Program<'info, Token>,
-    pub associated_token_program: Program<'info, AssociatedToken>,
-    pub rent: Sysvar<'info, Rent>,
-}
-
-#[derive(Accounts)]
-#[instruction(state_in: StateInput)]
-pub struct UpdateState<'info> {
-    #[account(
-        mut,
-        has_one = update_authority
-    )]
-    pub state: Account<'info, State>,
-
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    pub update_authority: Signer<'info>,
-
-    #[account()]
-    pub msol_mint: Box<Account<'info, Mint>>,
-    #[account()]
-    pub bsol_mint: Box<Account<'info, Mint>>,
-
-    /// Must be a PDA, but otherwise owned by the system account ie not initialised with data
-    #[account(
-    seeds = [state.key().as_ref(), MSOL_ACCOUNT],
-    bump = state_in.msol_authority_bump
-    )]
-    pub msol_token_account_authority: SystemAccount<'info>,
-
-    #[account(mut)]
-    /// CHECK: If owned by the system program, it is created as an ATA
-    pub msol_token_account: UncheckedAccount<'info>,
-
-    #[account()]
-    pub liq_pool_mint: Box<Account<'info, Mint>>,
-
-    #[account(mut)]
-    /// CHECK: If owned by the system program, it is created as an ATA
-    pub liq_pool_token_account: UncheckedAccount<'info>,
-
-    #[account(
-    seeds = [state.key().as_ref(), BSOL_ACCOUNT],
-    bump = state_in.bsol_authority_bump
-    )]
-    pub bsol_token_account_authority: SystemAccount<'info>,
-
-    #[account(mut)]
-    /// CHECK: If owned by the system program, it is created as an ATA
-    pub bsol_token_account: UncheckedAccount<'info>,
-
-    pub system_program: Program<'info, System>,
-    pub token_program: Program<'info, Token>,
-    pub associated_token_program: Program<'info, AssociatedToken>,
-    pub rent: Sysvar<'info, Rent>,
-}
-
-#[derive(Accounts)]
-#[instruction(size: usize)]
-pub struct ResizeState<'info> {
-    #[account(
-    mut,
-    has_one = update_authority,
-    realloc = size,
-    realloc::payer = payer,
-    realloc::zero = false,
-    )]
-    pub state: Account<'info, State>,
-
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    pub update_authority: Signer<'info>,
-
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts, Clone)]
-pub struct Deposit<'info> {
-    #[account(mut,has_one = marinade_state)]
-    pub state: Box<Account<'info, State>>,
-
-    #[account(mut)]
-    pub marinade_state: Box<Account<'info, MarinadeState>>,
-
-    #[account(
-    mut,
-    constraint = gsol_mint.mint_authority == COption::Some(gsol_mint_authority.key()),
-    )]
-    pub gsol_mint: Box<Account<'info, Mint>>,
-
-    #[account(
-    seeds = [
-    state.key().as_ref(),
-    GSOL_MINT_AUTHORITY,
-    ],
-    bump = state.gsol_mint_authority_bump,
-    )]
-    pub gsol_mint_authority: SystemAccount<'info>,
-
-    #[account(mut)]
-    pub msol_mint: Box<Account<'info, Mint>>,
-
-    #[account(mut)]
-    pub liq_pool_mint: Box<Account<'info, Mint>>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub liq_pool_sol_leg_pda: AccountInfo<'info>,
-
-    #[account(mut)]
-    pub liq_pool_msol_leg: Box<Account<'info, TokenAccount>>,
-    /// CHECK: Checked in marinade program
-    pub liq_pool_msol_leg_authority: AccountInfo<'info>,
-
-    /// CHECK: Checked in marinade program
-    pub liq_pool_mint_authority: AccountInfo<'info>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub reserve_pda: AccountInfo<'info>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub transfer_from: Signer<'info>,
-
-    #[account(
-    mut,
-    token::mint = msol_mint,
-    token::authority = msol_token_account_authority,
-    )]
-    pub mint_msol_to: Account<'info, TokenAccount>,
-
-    #[account(
-    mut,
-    token::mint = liq_pool_mint,
-    token::authority = msol_token_account_authority,
-    )]
-    pub mint_liq_pool_to: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-    mut,
-    token::mint = gsol_mint,
-    token::authority = transfer_from.key(),
-    )]
-    pub mint_gsol_to: Account<'info, TokenAccount>,
-
-    /// CHECK: Checked in marinade program
-    pub msol_mint_authority: AccountInfo<'info>,
-
-    #[account(
-    seeds = [state.key().as_ref(), MSOL_ACCOUNT],
-    bump = state.msol_authority_bump
-    )]
-    pub msol_token_account_authority: SystemAccount<'info>,
-
-    pub system_program: Program<'info, System>,
-    pub token_program: Program<'info, Token>,
-    pub marinade_program: Program<'info, MarinadeFinance>,
-}
-
-#[derive(Accounts, Clone)]
-pub struct DepositStakeAccount<'info> {
-    #[account(mut, has_one = marinade_state)]
-    pub state: Box<Account<'info, State>>,
-
-    #[account(mut)]
-    pub marinade_state: Box<Account<'info, MarinadeState>>,
-
-    #[account(
-        mut,
-        constraint = gsol_mint.mint_authority == COption::Some(gsol_mint_authority.key()),
-    )]
-    pub gsol_mint: Box<Account<'info, Mint>>,
-
-    #[account(
-    seeds = [
-    state.key().as_ref(),
-    GSOL_MINT_AUTHORITY,
-    ],
-    bump = state.gsol_mint_authority_bump,
-    )]
-    pub gsol_mint_authority: SystemAccount<'info>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub validator_list: AccountInfo<'info>,
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub stake_list: AccountInfo<'info>,
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub stake_account: AccountInfo<'info>,
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub duplication_flag: AccountInfo<'info>,
-
-    /// Marinade makes a distinction between the `stake_authority`(proof of ownership of stake account)
-    /// and the `rent_payer`(pays to init the validator_record account). Both are required to be signers
-    /// for the instruction. These two accounts can be treated as one and the same, and here, they are.
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub stake_authority: Signer<'info>,
-
-    #[account(mut)]
-    pub msol_mint: Box<Account<'info, Mint>>,
-
-    #[account(
-    mut,
-    token::mint = msol_mint,
-    token::authority = msol_token_account_authority,
-    )]
-    pub mint_msol_to: Account<'info, TokenAccount>,
-
-    #[account(
-    mut,
-    token::mint = gsol_mint,
-    token::authority = stake_authority.key(),
-    )]
-    pub mint_gsol_to: Account<'info, TokenAccount>,
-
-    /// CHECK: Checked in marinade program
-    pub msol_mint_authority: AccountInfo<'info>,
-
-    #[account(
-    seeds = [state.key().as_ref(), MSOL_ACCOUNT],
-    bump = state.msol_authority_bump
-    )]
-    pub msol_token_account_authority: SystemAccount<'info>,
-
-    pub clock: Sysvar<'info, Clock>,
-    pub rent: Sysvar<'info, Rent>,
-
-    /// CHECK: Checked in marinade program
-    pub stake_program: AccountInfo<'info>,
-    pub system_program: Program<'info, System>,
-    pub token_program: Program<'info, Token>,
-    pub marinade_program: Program<'info, MarinadeFinance>,
-}
-
-#[derive(Accounts, Clone)]
-#[instruction(lamports: u64)]
-pub struct LiquidUnstake<'info> {
-    #[account(
-    has_one = marinade_state,
-    )]
-    pub state: Box<Account<'info, State>>,
-
-    #[account(mut)]
-    pub marinade_state: Box<Account<'info, MarinadeState>>,
-
-    #[account(mut)]
-    pub msol_mint: Box<Account<'info, Mint>>,
-
-    #[account(mut)]
-    pub liq_pool_mint: Box<Account<'info, Mint>>,
-
-    #[account(
-    mut,
-    constraint = gsol_mint.mint_authority == COption::Some(gsol_mint_authority.key()),
-    )]
-    pub gsol_mint: Box<Account<'info, Mint>>,
-
-    #[account(
-    seeds = [
-    state.key().as_ref(),
-    GSOL_MINT_AUTHORITY,
-    ],
-    bump = state.gsol_mint_authority_bump,
-    )]
-    /// Used to ensure the correct GSOL mint is used
-    pub gsol_mint_authority: SystemAccount<'info>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub liq_pool_sol_leg_pda: AccountInfo<'info>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub liq_pool_msol_leg: Box<Account<'info, TokenAccount>>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub liq_pool_msol_leg_authority: SystemAccount<'info>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub treasury_msol_account: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-    mut,
-    token::mint = msol_mint,
-    token::authority = get_msol_from_authority,
-    )]
-    pub get_msol_from: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-    seeds = [state.key().as_ref(), MSOL_ACCOUNT],
-    bump = state.msol_authority_bump
-    )]
-    pub get_msol_from_authority: SystemAccount<'info>, // sunrise-stake PDA
-
-    #[account(
-    mut,
-    token::mint = liq_pool_mint,
-    // use the same authority PDA for this and the msol token account
-    token::authority = get_msol_from_authority
-    )]
-    pub get_liq_pool_token_from: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-    mut,
-    token::authority = gsol_token_account_authority
-    )]
-    pub gsol_token_account: Box<Account<'info, TokenAccount>>,
-
-    #[account(mut)]
-    /// CHECK: Owner of the gsol token account
-    pub gsol_token_account_authority: Signer<'info>,
-
-    pub rent: Sysvar<'info, Rent>,
-    pub system_program: Program<'info, System>,
-    pub token_program: Program<'info, Token>,
-    pub marinade_program: Program<'info, MarinadeFinance>,
-}
-
-#[derive(Accounts, Clone)]
-#[instruction(
-epoch: u64,
-order_unstake_ticket_index: u64,
-order_unstake_ticket_account_bump: u8,
-previous_order_unstake_ticket_management_account_bump: u8,
-)]
-pub struct TriggerPoolRebalance<'info> {
-    #[account(
-    has_one = marinade_state,
-    has_one = gsol_mint,
-    )]
-    pub state: Box<Account<'info, State>>,
-
-    #[account(mut)]
-    pub marinade_state: Box<Account<'info, MarinadeState>>,
-
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    #[account(mut)]
-    pub msol_mint: Box<Account<'info, Mint>>,
-
-    pub gsol_mint: Box<Account<'info, Mint>>,
-
-    #[account(mut)]
-    pub liq_pool_mint: Box<Account<'info, Mint>>,
-
-    /// CHECK: Checked in marinade program
-    pub liq_pool_mint_authority: AccountInfo<'info>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub liq_pool_sol_leg_pda: AccountInfo<'info>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub liq_pool_msol_leg: Box<Account<'info, TokenAccount>>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub liq_pool_msol_leg_authority: SystemAccount<'info>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub treasury_msol_account: AccountInfo<'info>,
-
-    #[account(
-    mut,
-    token::mint = msol_mint,
-    token::authority = get_msol_from_authority,
-    )]
-    pub get_msol_from: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-    mut,
-    seeds = [state.key().as_ref(), MSOL_ACCOUNT],
-    bump = state.msol_authority_bump
-    )]
-    pub get_msol_from_authority: SystemAccount<'info>, // sunrise-stake PDA
-
-    #[account(
-    mut,
-    token::mint = liq_pool_mint,
-    // use the same authority PDA for this and the msol token account
-    token::authority = get_msol_from_authority
-    )]
-    pub liq_pool_token_account: Box<Account<'info, TokenAccount>>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub reserve_pda: UncheckedAccount<'info>,
-
-    #[account(
-    mut,
-    // This will be initialised programmatically if it is needed
-    // init,
-    // payer = gsol_token_account_authority,
-    // space = MARINADE_TICKET_ACCOUNT_SPACE,
-    seeds = [state.key().as_ref(), ORDER_UNSTAKE_TICKET_ACCOUNT, &epoch.to_be_bytes(), &order_unstake_ticket_index.to_be_bytes()],
-    bump = order_unstake_ticket_account_bump,
-    )]
-    /// CHECK: Checked in marinade program
-    pub order_unstake_ticket_account: UncheckedAccount<'info>,
-
-    #[account(
-    init_if_needed,
-    space = OrderUnstakeTicketManagementAccount::SPACE,
-    payer = payer,
-    seeds = [state.key().as_ref(), ORDER_UNSTAKE_TICKET_MANAGEMENT_ACCOUNT, &epoch.to_be_bytes()],
-    bump
-    )]
-    pub order_unstake_ticket_management_account:
-        Box<Account<'info, OrderUnstakeTicketManagementAccount>>,
-
-    #[account(
-    mut,
-    // TODO - before closing, we need to be sure all order tickets are closed
-    // one way to do this is to decrement the number of open tickets for each one chaimed in this tx
-    // and fail out if there are any left
-    // This works until we have too many open tickets that it needs more than one tx
-    // close = payer,
-    seeds = [state.key().as_ref(), ORDER_UNSTAKE_TICKET_MANAGEMENT_ACCOUNT, &(epoch - 1).to_be_bytes()],
-    bump = previous_order_unstake_ticket_management_account_bump
-    )]
-    pub previous_order_unstake_ticket_management_account:
-        Option<Account<'info, OrderUnstakeTicketManagementAccount>>,
-
-    pub clock: Sysvar<'info, Clock>,
-    pub rent: Sysvar<'info, Rent>,
-    pub system_program: Program<'info, System>,
-    pub token_program: Program<'info, Token>,
-    pub marinade_program: Program<'info, MarinadeFinance>,
-}
-
-#[derive(Accounts, Clone)]
-pub struct OrderUnstake<'info> {
-    #[account(
-    has_one = treasury,
-    has_one = marinade_state,
-    )]
-    pub state: Box<Account<'info, State>>,
-
-    #[account()]
-    pub marinade_state: Box<Account<'info, MarinadeState>>,
-
-    #[account(mut)]
-    pub msol_mint: Account<'info, Mint>,
-
-    #[account(
-    constraint = gsol_mint.mint_authority == COption::Some(gsol_mint_authority.key()),
-    )]
-    pub gsol_mint: Box<Account<'info, Mint>>,
-
-    #[account(
-    seeds = [
-    state.key().as_ref(),
-    GSOL_MINT_AUTHORITY,
-    ],
-    bump = state.gsol_mint_authority_bump,
-    )]
-    /// Used to ensure the correct GSOL mint is used
-    pub gsol_mint_authority: SystemAccount<'info>,
-
-    #[account(
-    mut,
-    token::mint = msol_mint,
-    token::authority = get_msol_from_authority,
-    )]
-    pub get_msol_from: Account<'info, TokenAccount>,
-
-    #[account(
-    seeds = [state.key().as_ref(), MSOL_ACCOUNT],
-    bump = state.msol_authority_bump
-    )]
-    pub get_msol_from_authority: SystemAccount<'info>, // sunrise-stake PDA
-
-    #[account(
-    mut,
-    token::authority = gsol_token_account_authority
-    )]
-    pub gsol_token_account: Account<'info, TokenAccount>,
-
-    #[account(mut)]
-    /// Owner of the gSOL
-    pub gsol_token_account_authority: Signer<'info>,
-
-    #[account(zero, rent_exempt = enforce)]
-    /// CHECK: Checked in marinade program
-    pub new_ticket_account: UncheckedAccount<'info>,
-
-    #[account(init, space = SunriseTicketAccount::SPACE, payer = gsol_token_account_authority)]
-    pub sunrise_ticket_account: Account<'info, SunriseTicketAccount>,
-
-    #[account()]
-    /// CHECK: Matches state.treasury
-    pub treasury: SystemAccount<'info>, // sunrise-stake treasury
-
-    pub clock: Sysvar<'info, Clock>,
-    pub rent: Sysvar<'info, Rent>,
-    pub token_program: Program<'info, Token>,
-    pub marinade_program: Program<'info, MarinadeFinance>,
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts, Clone)]
-pub struct ClaimUnstakeTicket<'info> {
-    #[account(
-    has_one = marinade_state,
-    )]
-    pub state: Box<Account<'info, State>>,
-    #[account(mut)]
-    pub marinade_state: Box<Account<'info, MarinadeState>>,
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub reserve_pda: UncheckedAccount<'info>,
-
-    #[account(mut)]
-    pub marinade_ticket_account: Account<'info, MarinadeTicketAccount>,
-
-    #[account(
-        mut,
-        close = transfer_sol_to,
-        has_one = marinade_ticket_account,
-        constraint = sunrise_ticket_account.beneficiary == transfer_sol_to.key(),
-    )]
-    pub sunrise_ticket_account: Account<'info, SunriseTicketAccount>,
-
-    #[account(
-    mut,
-    seeds = [state.key().as_ref(), MSOL_ACCOUNT],
-    bump = state.msol_authority_bump
-    )]
-    pub msol_authority: SystemAccount<'info>, // sunrise-stake PDA
-
-    #[account(mut)]
-    pub transfer_sol_to: Signer<'info>,
-
-    pub clock: Sysvar<'info, Clock>,
-
-    pub marinade_program: Program<'info, MarinadeFinance>,
-
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts, Clone)]
-pub struct ExtractToTreasury<'info> {
-    #[account(
-    has_one = treasury,
-    has_one = marinade_state,
-    has_one = blaze_state,
-    )]
-    pub state: Box<Account<'info, State>>,
-
-    #[account(mut)]
-    pub marinade_state: Box<Account<'info, MarinadeState>>,
-
-    /// CHECK: Checked
-    pub blaze_state: AccountInfo<'info>,
-
-    #[account(mut)]
-    pub msol_mint: Box<Account<'info, Mint>>,
-
-    #[account()]
-    pub gsol_mint: Box<Account<'info, Mint>>,
-
-    pub bsol_mint: Box<Account<'info, Mint>>,
-
-    #[account()]
-    pub liq_pool_mint: Box<Account<'info, Mint>>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub liq_pool_sol_leg_pda: AccountInfo<'info>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub liq_pool_msol_leg: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-    mut,
-    token::mint = liq_pool_mint,
-    // use the same authority PDA for this and the msol token account
-    token::authority = get_msol_from_authority
-    )]
-    pub liq_pool_token_account: Box<Account<'info, TokenAccount>>,
-
-    #[account(mut)]
-    /// CHECK: Checked in marinade program
-    pub treasury_msol_account: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-    mut,
-    token::mint = msol_mint,
-    token::authority = get_msol_from_authority,
-    )]
-    pub get_msol_from: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-    seeds = [state.key().as_ref(), MSOL_ACCOUNT],
-    bump = state.msol_authority_bump
-    )]
-    pub get_msol_from_authority: SystemAccount<'info>, // sunrise-stake PDA
-
-    #[account(
-    mut,
-    token::mint = bsol_mint,
-    token::authority = get_bsol_from_authority,
-    )]
-    pub get_bsol_from: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-    seeds = [state.key().as_ref(), BSOL_ACCOUNT],
-    bump = state.bsol_authority_bump
-    )]
-    pub get_bsol_from_authority: SystemAccount<'info>, // sunrise-stake PDA
-
-    #[account(mut)]
-    /// CHECK: Matches state.treasury
-    pub treasury: SystemAccount<'info>, // sunrise-stake treasury
-
-    pub system_program: Program<'info, System>,
-    pub token_program: Program<'info, Token>,
-    pub marinade_program: Program<'info, MarinadeFinance>,
-}
-
-#[derive(Accounts, Clone)]
-pub struct CreateMetadata<'info> {
-    #[account(
-    has_one = marinade_state,
-    has_one = update_authority,
-    )]
-    pub state: Box<Account<'info, State>>,
-
-    #[account()]
-    pub marinade_state: Box<Account<'info, MarinadeState>>,
-
-    #[account(
-    mut,
-    constraint = gsol_mint.mint_authority == COption::Some(gsol_mint_authority.key()),
-    )]
-    pub gsol_mint: Box<Account<'info, Mint>>,
-
-    #[account(
-      seeds = [
-      state.key().as_ref(),
-      GSOL_MINT_AUTHORITY,
-      ],
-      bump = state.gsol_mint_authority_bump,
-      )]
-    pub gsol_mint_authority: SystemAccount<'info>,
-
-    pub update_authority: Signer<'info>,
-
-    pub system_program: Program<'info, System>,
-    pub rent: Sysvar<'info, Rent>,
-    pub token_program: Program<'info, Token>,
-
-    /// CHECK:
-    #[account(mut)]
-    pub metadata: AccountInfo<'info>,
-    /// CHECK:
-    pub token_metadata_program: AccountInfo<'info>,
-}
-
-#[derive(Accounts, Clone)]
-pub struct UpdateMetadata<'info> {
-    #[account(
-    has_one = marinade_state,
-    has_one = update_authority,
-    )]
-    pub state: Box<Account<'info, State>>,
-
-    #[account()]
-    pub marinade_state: Box<Account<'info, MarinadeState>>,
-
-    #[account(
-    mut,
-    constraint = gsol_mint.mint_authority == COption::Some(gsol_mint_authority.key()),
-    )]
-    pub gsol_mint: Box<Account<'info, Mint>>,
-
-    #[account(
-      seeds = [
-      state.key().as_ref(),
-      GSOL_MINT_AUTHORITY,
-      ],
-      bump = state.gsol_mint_authority_bump,
-      )]
-    pub gsol_mint_authority: SystemAccount<'info>,
-
-    pub update_authority: Signer<'info>,
-
-    /// CHECK:
-    #[account(mut)]
-    pub metadata: AccountInfo<'info>,
-    /// CHECK:
-    pub token_metadata_program: AccountInfo<'info>,
-}
-
-// If imported from the marinade crate, deserialisation does not work
-// TODO fix
-#[derive(Debug, BorshDeserialize)]
-pub struct TicketAccountData {
-    pub discriminator: u64,
-    pub state_address: Pubkey, // instance of marinade state this ticket belongs to
-    pub beneficiary: Pubkey,   // main account where to send SOL when claimed
-    pub lamports_amount: u64,  // amount this ticked is worth
-    pub created_epoch: u64, // epoch when this acc was created (epoch when delayed-unstake was requested)
-}
-
-#[error_code]
-pub enum ErrorCode {
-    #[msg("An error occurred when calculating an MSol value")]
-    CalculationFailure,
-    #[msg("Stake account deposit must be delegated")]
-    NotDelegated,
-    #[msg("Wrong update authority for Sunrise state")]
-    InvalidUpdateAuthority,
-    #[msg("Invalid Program Account")]
-    InvalidProgramAccount,
-    #[msg("Invalid Mint")]
-    InvalidMint,
-    #[msg("Unexpected Accounts")]
-    UnexpectedAccounts,
-    #[msg("Unexpected gsol mint supply")]
-    UnexpectedMintSupply,
-    #[msg("The order unstake management account is invalid for this epoch")]
-    InvalidOrderUnstakeManagementAccount,
 }
