+++ programs/sunrise-stake/src/state.rs
@@ -1,0 +1,176 @@
+use crate::error::ErrorCode;
+use crate::utils::seeds::GSOL_MINT_AUTHORITY;
+use anchor_lang::prelude::borsh::BorshDeserialize;
+use anchor_lang::prelude::*;
+use anchor_lang::solana_program::program_option::COption;
+use anchor_spl::token::{Mint, Token};
+use marinade_cpi::State as MarinadeState;
+
+#[account]
+pub struct State {
+    pub marinade_state: Pubkey,
+
+    pub update_authority: Pubkey,
+    pub gsol_mint: Pubkey,
+    pub treasury: Pubkey,
+    pub gsol_mint_authority_bump: u8,
+    pub msol_authority_bump: u8,
+
+    /// 0-100 - The proportion of the total staked SOL that should be in the
+    /// liquidity pool.
+    pub liq_pool_proportion: u8,
+    /// 0-100 - If unstaking would result in the proportion of SOL in the
+    /// liquidity pool dropping below this value, trigger an delayed unstake
+    /// for the difference
+    pub liq_pool_min_proportion: u8,
+
+    pub blaze_state: Pubkey,
+    pub marinade_minted_gsol: u64,
+    pub blaze_minted_gsol: u64,
+    pub bsol_authority_bump: u8,
+}
+
+impl State {
+    pub const SPACE: usize = 32 + 32 + 32 + 32 + 1 + 1 + 1 + 1 + 32 + 8 + 8 + 1 + 8 /* DISCRIMINATOR */ ;
+
+    pub fn set_values(
+        &mut self,
+        input: &StateInput,
+        gsol_mint: &Pubkey,
+        marinade_minted_gsol: u64,
+        blaze_minted_gsol: u64,
+    ) {
+        self.marinade_state = input.marinade_state;
+        self.blaze_state = input.blaze_state;
+        self.update_authority = input.update_authority;
+        self.gsol_mint_authority_bump = input.gsol_mint_authority_bump;
+        self.msol_authority_bump = input.msol_authority_bump;
+        self.bsol_authority_bump = input.bsol_authority_bump;
+        self.treasury = input.treasury;
+        self.gsol_mint = *gsol_mint;
+        self.liq_pool_proportion = input.liq_pool_proportion;
+        self.liq_pool_min_proportion = input.liq_pool_min_proportion;
+        self.marinade_minted_gsol = marinade_minted_gsol;
+        self.blaze_minted_gsol = blaze_minted_gsol;
+    }
+}
+
+#[derive(AnchorSerialize, AnchorDeserialize, Clone)]
+pub struct StateInput {
+    pub marinade_state: Pubkey,
+    pub blaze_state: Pubkey,
+    pub update_authority: Pubkey,
+    pub treasury: Pubkey,
+    pub gsol_mint_authority_bump: u8,
+    pub msol_authority_bump: u8,
+    pub bsol_authority_bump: u8,
+    pub liq_pool_proportion: u8,
+    pub liq_pool_min_proportion: u8,
+}
+
+/// Maps a marinade ticket account to a GSOL token holder
+#[account]
+pub struct SunriseTicketAccount {
+    pub state_address: Pubkey, // instance of sunrise state this ticket belongs to
+    pub marinade_ticket_account: Pubkey,
+    pub beneficiary: Pubkey,
+}
+impl SunriseTicketAccount {
+    pub const SPACE: usize = 32 + 32 + 32 + 8 /* DISCRIMINATOR */ ;
+}
+
+#[account]
+pub struct OrderUnstakeTicketManagementAccount {
+    pub state_address: Pubkey,
+    pub epoch: u64,
+    pub tickets: u64,
+    pub total_ordered_lamports: u64,
+}
+impl OrderUnstakeTicketManagementAccount {
+    pub const SPACE: usize = 32 + 8 + 8 + 8 + 8 /* DISCRIMINATOR */ ;
+
+    pub fn add_ticket(&mut self, ticket_amount_lamports: u64, epoch: u64) -> Result<()> {
+        if self.epoch == 0 {
+            self.epoch = epoch;
+        } else {
+            require_eq!(
+                self.epoch,
+                epoch,
+                ErrorCode::InvalidOrderUnstakeManagementAccount
+            );
+        }
+        self.tickets += 1;
+        self.total_ordered_lamports += ticket_amount_lamports;
+        Ok(())
+    }
+}
+
+// If imported from the marinade crate, deserialisation does not work
+// TODO fix
+#[derive(Debug, BorshDeserialize)]
+pub struct TicketAccountData {
+    pub discriminator: u64,
+    pub state_address: Pubkey, // instance of marinade state this ticket belongs to
+    pub beneficiary: Pubkey,   // main account where to send SOL when claimed
+    pub lamports_amount: u64,  // amount this ticked is worth
+    pub created_epoch: u64, // epoch when this acc was created (epoch when delayed-unstake was requested)
+}
+
+#[derive(Accounts, Clone)]
+pub struct CreateMetadata<'info> {
+    #[account(
+    has_one = marinade_state,
+    has_one = update_authority,
+    )]
+    pub state: Box<Account<'info, State>>,
+
+    #[account()]
+    pub marinade_state: Box<Account<'info, MarinadeState>>,
+
+    #[account(
+    mut,
+    constraint = gsol_mint.mint_authority == COption::Some(gsol_mint_authority.key()),
+    )]
+    pub gsol_mint: Box<Account<'info, Mint>>,
+
+    #[account(
+      seeds = [
+      state.key().as_ref(),
+      GSOL_MINT_AUTHORITY,
+      ],
+      bump = state.gsol_mint_authority_bump,
+      )]
+    pub gsol_mint_authority: SystemAccount<'info>,
+
+    pub update_authority: Signer<'info>,
+
+    pub system_program: Program<'info, System>,
+    pub rent: Sysvar<'info, Rent>,
+    pub token_program: Program<'info, Token>,
+
+    /// CHECK:
+    #[account(mut)]
+    pub metadata: AccountInfo<'info>,
+    /// CHECK:
+    pub token_metadata_program: AccountInfo<'info>,
+}
+
+#[derive(Accounts)]
+#[instruction(size: usize)]
+pub struct ResizeState<'info> {
+    #[account(
+    mut,
+    has_one = update_authority,
+    realloc = size,
+    realloc::payer = payer,
+    realloc::zero = false,
+    )]
+    pub state: Account<'info, State>,
+
+    #[account(mut)]
+    pub payer: Signer<'info>,
+
+    pub update_authority: Signer<'info>,
+
+    pub system_program: Program<'info, System>,
+}
